+++ packages/terra-core-docs/src/terra-dev-site/test/section-header/StickyTitleSectionHeader.test.jsx
@@ -1,0 +1,29 @@
+/* eslint-disable react/forbid-dom-props */
+import React from 'react';
+import SectionHeader from 'terra-section-header';
+
+export default () => (
+  <div style={{
+    width: '400px', height: '200px', background: 'gray', margin: '10px', overflow: 'scroll',
+  }}
+  >
+    <div style={{ width: '600px', height: '400px' }}>
+      <SectionHeader
+        text="Closed Section Header 1"
+        isTitleSticky
+        onClick={() => {}}
+      />
+      <br />
+      <SectionHeader
+        text="Closed Section Header 2"
+        isTitleSticky
+        onClick={() => {}}
+      />
+      <br />
+      <SectionHeader
+        text="Section Header 3"
+        isTitleSticky
+      />
+    </div>
+  </div>
+);